Switch the old thunderbird*.in files to icedove.*
By removing the the applications name switch in the rules file the old
debian/*.in files are not needed any longer. We only suppport building
Icedove related packages.

Depending on this we also don't need to take care about application
names in the menu or post and preinst scripts. Remove all old
substitutions variables inside the files.