debian/rules: add override for dh_install
calendar-google-provider and iceowl-extension need special handling, we
repack the already created *.xpi files later. So we have to extract them
first.

Because of the branding we also have to install our own icons for the
Icedove grafics.

The icedove-dev package doesn't needs to ship the
lib{mozalloc,mozjs,xul}.so. We link them to the librarys which come
with the standard icedove package. Removed for deduplication.