debian/control: remove build-dep on libnotify-dev
Mozilla isn't using the notify mechanismus of new messages etc. via
message notify daemon since several versions so we don't ned this build
depenency anymore.
Note: Ubuntu is using the old message notify function because they
maintain a patch to provide this functionality. In Debian we are unably
to do this too do a lack of free time to maintain such a patch too on
top of the current patchqueue.