debian/rules: adding switch for no icedove-dbg build
While testing various build cases of Icedove the build of the
icedove-dbg package will take much time, but mostly we don't ever use
the -dbg package.

The waiting time can be significant reduced if we skip the build of the
-dbg package while developing a new version or fixing something.

This commit implements an 'if then else' construct that check if the
debian/changelog has a "Release == UNRELEASED" or not.
In such a case the build of the -dbg package will be skipped, otherwise
it will be like all the other packages build. This ensures a normal
build of all packages on the autobuilders