fixup branding about.png file
With version 42.0~b1 cames GTK3 support along. If you use the dark theme
in Gnome there is a small part in the bottom of about.png visible that
has a transparent part. By this the About page after starting Icedove
looks ugly, because the bottom part of the graphic shows the dark
background and the version string on top of the graphic isn't really
readable. Fixing the PNG fixes this behaviour.