icedove-l10n-de: replace recommends for myspell-de-*
The Recommends for the myspell-de-* packages are repleaced by the
respective hunspell packages. The myspell package are dropped.

The myspell and hunspell dictionaries are build from the same source,
with some functional extensions of the huspell dictionaries, so the
hunspell dictionaries should always be preferred over the myspell ones
(if the calling program supports hunspell).

Closes: #813830