Icedove Branding: adopt usptream changes to branding
Upstream changed the makefile magic around the branding so we need to
adopt this within the files

  debian/icedove-branding/Makefile.in
  debian/icedove-branding/jar.mn

And on doing changes to the Debian specific branding also adopting the
license changes done in the past to the respective files which
originally taken from upstream.
All files now licensend under MPL2.0.