adding helper script create-icedove-l10n-tarball.sh
For the various l10n packages we need the helper script which was
already used in the icedove-l10n source package to create a orig.tar.xz
file with the source auf the l10n packages.
This script is currently only prepared to pull the files for a Beta
release of the Thunderbird.