adding helper script create-iceowl-l10n-tarball.sh
As for the Icedove l10n files the language files for Iceowl (aka
Lightning) need to be collected as a subtarball with a naming
icedove_VERSION.orig-iceowl-l10n.tar.xz.
The script was existing in the SRC package of iceowl-l10n and was
slightly modified to create the correct named tarball for example. Like
before for create-icedove-l10n-tarball.sh the script is currently only
tested for beta releases of Lightning.