de-branding: adopting dh files for icedove-dev package
Because of the changes in previous commits the appname is changed (back)
to 'thunderbird'.
The build system is using this variable on various targets. Due this the
reflecting binaries created while the build is changing too.
This needs a adjusting of the installation and the symlinking inside the
Debian packaging for the icedove-dev package.

The folder /etc/icedove will change later, also the changes to the pre|post
scripts.