de-branding: (re)adding a wrapper script for TB starting
Because of the de-branding the profile for Thunderbird has changed back
to the Mozilla default $HOME/.thunderbird.

Related to the old branded binary name Icedove the profile folder
$HOME/.icedove was previously used. To get a uninterrupted workflow for
the user while going back to a Thunderbird package we have to take care
about the profile migration to the new profile folder.

This commit creates a wrapper script that will do this migration of the
user data. It's purpose is to act as a wrapper while calling
/usr/lib/thunderbird/thunderbird through /user/bin/thunderbird.
It will also handle a easier handling for creating gdb logs if needed.
Other debuggers can be added to but not yet or probably never. Right now
we never had the need to provide a valgrind build of Icedove or
Thunderbird. Any help is appreciated here.