thunderbird: switching dpkg-maintscript-helper to *.maintscript
Using the maintscript mechanism from dpkg for the thunderbird and
icedove package. Due this also remove the old thunderbird.{preinst,postrm}
file and the dpkg-maintscript-helper part from thunderbird.postinst as
the old mentioned version 5.0 was valid for Icedove and not Thunderbird.

The postinst for Thunderbird is still needed as we have to move possible
other files than icedove.js from

  /etc/icedove/pref/ --> /etc/thunderbird/pref

Icedove 5.0 was alive within the Squeeze (Debian 6.0) release, which is
gone since several weeks from now.