de-branding: let helper scripts reflect thunderbird change
After the remove the Debian branding we need to rework the the helper
scripts that downloading the l10n files for Thunderbird and Lightning to
resepct the name changes and create respective subtarball archives.

Note!
After this change debian/rules need to be changed too after a import of
new folders due the changed directory names for the l10n stuff!