debian/icedove.js: disable Icedove startup check
Without the setting 'pref("mail.shell.checkDefaultClient", false);'
Thunderbird will ask on first start if the user would like to register
Icedove at least as default email application.

If the user is answering this question by clicking "set as default" an
new desktop file will be created in

  $HOME/.local/share/applications/userapp-Icedove-[xxxxxx].desktop

and various new associations in

  ~/.config/mimeapps.list
and
  ~/.local/share/applications/mimeapps.list

will be created.

Due this the system is finding probably two valid extries for
'Name=Icedove' and is adding those application into various menus.

Closes: #817973