debian/rules: adding additional CFLAGS and CXXFLAGS
Some source files checking pointer aginst NULL but those are supposed to
be allways valid and the GCC-6 compiler is optimizing those checks out,
in the end the binary got some nullpointer and is segfaulting.
As a workaround say the compiler not to optimize those checks so there
are no nullpointer exceptions.

This commit is a needed addition for
d87a18670c9e45ee05205f74814364b85a1f6c32

Please note also the refering upstream bug 1251576
https://bugzilla.mozilla.org/show_bug.cgi?id=1251576