debian/control: addjust breaks for xul-ext-foxyproxy-standard
Until now we use a Breaks >> 3.4-1 for xul-ext-foxyproxy-standard,
related to this a user can't install this plugin together with Icedove.

This was done because introducing segfaults by the extension on versions
greater 3.4, David Prévot prepared a new upload of the
xul-ext-foxyproxy-standard extension wich is working again with Icedove,
adjusting the dependencies.

Closes: #825749