debian/icedove.js: disabling baselinejit functionality
Icedove is again already crashing as a result of segfaulting of the JS
engine. We already had reports abut this: #833864, #833591

The origin of this issues are mostly pointer into non initialised data
or deleted pointer probably in the JS engine. This change to the default
setting for Icedove is disabling the memory interaction beween different
JS actions on shared memory, aka the JIT Compiler [1].
Icedove will use more memory of course and will acting a bit slower, but
should not crash that much as without this change.

[1] https://blog.mozilla.org/javascript/2013/04/05/the-baseline-compiler-has-landed/

Closes: #837930