thunderbird-wrapper.sh: adding fixing inside mimeTypes.rdf
The file 'mimeTypes.rdf' may have registered /usr/bin/iceweasel for
various application like http and https. This is no problem as long as
the user isn't removing the transitional iceweasel package which is
providing a symlink from /usr/bin/iceweasel -> /usr/bin/firefox.

So this commit is fixing this by changing all old entries for
/usr/bin/iceweasel with /usr/bin/x-www-browser which is pointing to the
prefered browser. This is only done while the profile migration is done.
If a user isn't using this automatism and is moving the profile folder
by hand the modifying of 'mimeTypes.rdf' has to also done manually.

Thanks to Jens Reyer who figured this problem and the solution out!

Closes: #837516