thunderbird-wrapper.sh: also migrate mimeapps.list
It's also a good idea to change existing MIME associations with the
icedove desktop within $(HOME)/.config/mimeapps.list as Laurent
Bigonville pointed out.

Improving the script a bit more so it will not going further and try
things which can lead to a inconsist old profile folder depending on
ignoring the first error.

Closes: #850864