crashreporter: enable the reporter for thunderbird
Now we are back on the thunderbird package we can use the crashreporter
again, that gives the user the avilibility to report crashes and issues
directly against upstream. That should also help maintaining the package
as it decreases the burden on reporting bugs also upstream from reported
bugs in the Debian bug tracker.