tb-helper*: rework option parsing for wrapper script
Using the option seperator '--' in the thunderbird-wrapper script isn't
usable, it breaks other scripts and caller functions.
To not break other stuff removing the getopt functionality as we need to
collect the possible mixed up options given to the wrapper script and
split them of.

Thanks Guido Günther for getting the basic while loop for handling this.

Closes: #855872