tb-wrapper*: complete rework and moving over for symlinking
The two helper scripts got completely restructured to fit into the idea
we do a symlinking of .thunderbird to .icedove in the users home.
To goal this the various steps got break of into subfunctions that are
called from the main wrapper helper script /usr/bin/thunderbird.

Also adding some extra wrapper option to get the fixing for the MIME
part also available ouside the profile adption. This commit is made also
from contribution of Debian fellows.
Thanks go out to Jens Reyer who contributed not only patches but also
checking the script against the shellcheck tool and by doing the needed
testing of this whole stuff!

The logic for the adoption is hopefully simple.
If there is no folder or already symlinked ~/.thunderbird folder to
.icedove we can simply create the link to ~/.icedove.
If we check there is a existing symlink ~/.thunderbird and isn't pointing
to ~/.icedove we error out then.

Closes: #855265, #855391, #855501, #856490