tb-helper*: fix up that silly comments behind the if statement
A stupid error happen in the if statement while checking for existing
folders .icedove and .thunderbird doing work for commit
c9e5b708a2700185168bdc3ce2a7224b3f9a6404

The comment behind the if check is of course intepreted by the Bash as
we have a backslash '\' in the if condition to ignore the line break
here. By this the check wasn't working correctly and assumes the check
is failing and informs the user for a not valid folder situation.

Closes: #857029, #857032, #857098, #857112