tb-helper: pass arguments by using a array to TB call
The approach in c2a1d77dd811ffd758f8306a71629e28490bc628 wasn't the
right attempt, as explained by Simon and Daniel in #855334 the best
solution is to use a Bash array that's taken as argument for the TB
binary call.
It's not only more simple, it's also more secure than to use eval for
variable expansion, really bing thanks for explaining this potential
security problem goes out to Simon here! Really appreciated.
But also thanking Daniel for providing a patch with the needed
structural changes!